backend/requirements.txt
Flask
requests
google-generativeai
python-dotenv
Flask-Cors